made changes to isKekulean and implemented the testkekuleanThm function
The program can now test the two kekulean thms and will print out a
conflict between the thms to a text file.
The linear kekulean check is having issue in being implemented, so the
assignBonds function has been copied and modified to just find a single
Kekule structure. This is slower, but will work until the proper
algorithm is implemented.
There also does not seem to be an issue with the Fries/Clars numbers of
graphs; numbers reflected graphs today during testing.
The display graph method was changed so faces would have a cooler other
than white. This makes it possible to tell what is face and what is
not.