WIP on maximizing Clars numbers
Added neighbors to Face class. Graph class has new methods for
assigning the face neighbors and maximizing Clars numbers. Of note, the
neighbors attribute in Face does not seem to be updating.  In the
setter, it seems that neighbors does update with the correct number of
neighbors, but the getter is only returning an empty set.